pyproject.toml
[project]
name = "app"
version = "0.0.0"
dependencies = [
    "fastapi>=0.128.0,<0.129.0",
    "jder-fastapi==0.2.1,<0.3.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "scalar-fastapi>=1.6.1,<2.0.0",
    "uvicorn[standard]>=0.40.0,<0.41.0",
]
requires-python = ">=3.13.0,<4.0.0"

[dependency-groups]
dev = [
    "ruff>=0.14.14,<0.15.0",
    "ty>=0.0.14,<0.1.0",
]
